restrucutred code to use new endpoint
